pyproject.toml
[tool.poetry]
name = "five-nines-agent"
version = "0.1.0"
description = ""
authors = ["Sébastien Puyet <sebastien@fivenines.io>"]
license = "WTFPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.1"
psutil = "^5.9.8"
systemd-watchdog = "^0.9.0"

[tool.poetry.scripts]
my-script = "five_nines_agent.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
